limit the delay time for TimerSecs

+++ PLC_esp8266/main/LogicProgram/Inputs/TimerSecs.h
@@ -12,7 +12,7 @@ class TimerSecs : public TimerBase {
     const Bitmap *GetCurrentBitmap() override final;
 
   public:
-    TimerSecs(uint16_t delay_time_s, InputBase &prior_item);
+    TimerSecs(uint32_t delay_time_s, InputBase &prior_item);
     ~TimerSecs();
 
     bool DoAction() override final;
